Remove unnecessary FileSystem.exists check
Removes a pre-emptive FileSystem#exists(Path) check before attempting
to list path contents when absent partition errors are configured to
be ignored. Instead, ignoring absent partitions can be done as part
of a check only in the case where partition listing actually fails.

For file systems like S3, checking the existence of a "directory"
already incurs an S3 listing call, which are relatively expensive in
terms of API rate limits and latency.