Support more types of parameters to JSON path
Before this change, the JSON path parameters could be only
numeric, boolean, datetime and string values, and other values
that could be implicitly coerced to varchar.

After this change, any values are accepted which can be coerced
to varchar. It is no more required that an implicit coercion
is defined.