Drop no new enforced domains check in PushPredicateIntoTableScan
The check dates back to the time when `TupleDomain` was the only
information passed to the connector in the
`ConnectorMetadata.applyFilter`. Its purpose was to ensure the connector
does not erroneously return some new column constraints (`Domain`
objects) in `ConstraintApplicationResult.remainingFilter`.

With expression-based pushdown, the check is no longer valid. A
connector may be able to translate `Constraint.expression` (or part
thereof) into a `Domain` / `TupleDomain` and then enforce it, or return
such simplified representation as a remaining `TupleDomain`
(`ConstraintApplicationResult.remainingFilter`).